requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.0
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-mail-templated==2.6.5
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.6
flower==2.0.1
humanize==4.9.0
idna==3.4
inflection==0.5.1
kombu==5.2.4
oauthlib==3.2.2
packaging==23.1
prometheus_client==0.20.0
prompt-toolkit==3.0.38
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycurl==7.43.0.5
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
